feat: reads user secrets via container secrets
